Project layout refactor.
 + Moved independent files in self directories.
 + Small files naming modification for better association with inner methods.
