+++ src/distributionRout/DistributionRout.go
@@ -1,8 +1,7 @@
-package handler
+package distributionRout
 
 import (
 	"encoding/json"
-	"fmt"
 	"io"
 	"net/http"
 	"restaurant_client/src/utils"
@@ -10,6 +9,7 @@ import (
 
 func DistributionHandler(writer http.ResponseWriter, request *http.Request) {
 	var data utils.DistributionData
+	var response string
 
 	jsonData, err := io.ReadAll(request.Body)
 	if err != nil {
@@ -23,5 +23,5 @@ func DistributionHandler(writer http.ResponseWriter, request *http.Request) {
 		return
 	}
 
-	_, _ = fmt.Fprintf(writer, string(jsonData))
+	http.Error(writer, response, http.StatusOK)
 }
