+++ src/singleton/Singleton.go
